#22 [CHORE] mutation 함수 분리
